kafka-stream/requirements.txt
git+https://github.com/dpkp/kafka-python.git
kafka-python==2.0.3.dev0